Remove auto-linking in MSVC to sleef.lib because it may change name
The feature "EXPORT_LIBM" to be introduced will change the
file name of the library, when enabled.

Since the auto-linking is implemented as a #pragma inside
sleeflibm_header.h.org, which is not generated by CMake, it
is difficult to make the name of the auto-linked library change
depending on the CMake configuration.

Auto-linking is not required in most build systems that already
require the client program to list out all of its dependencies.
This comment includes CMake, and the libm-tester program here
that links to sleef.lib after it is built (on Windows).