Do not export DLL functions as stdcall on 32-bit Windows.
The calling convention was inconsistently specified as stdcall
on the exporting side (the DLL being compiled), and the
not overridden on the importing side (sleef.h, so the default
would be cdecl unless overridden in compiler settings),
so importing should not have worked anyway.

On 64-bit Windows, __stdcall has no effect, so this consistency
was not noticed.

The benefit to stdcall is marginal in this day and age:
it saves one x86 instruction for the caller to clean the stack
after calling a function.