Re-position VECTOR_CC macro in function declarations to work with MSVC
The calling convention must be specified in between the return
type and the name of the function, under the Microsoft C/C++ compiler.

This change is to prepare for setting the calling convention __vectorcall.
It was done mechanically using a sed script.

Fortunately GCC is flexible in how the function attribute for
the calling convention is placed, and accepts the modified syntax.