Do not export DLL functions as __stdcall on 32-bit Windows
The calling convention was inconsistently specified as stdcall
on the exporting side (the DLL being compiled), and then
not overridden on the importing side (sleef.h, so the default
would be cdecl unless overridden in compiler settings),
so importing would not have worked anyway.

On 64-bit Windows, __stdcall has no effect, so this consistency
was not noticed.

The benefit to stdcall is marginal in this day and age:
it saves one x86 instruction for the caller to clean the stack
after calling a function.

Finally there is a good reason to choose __cdecl over __stdcall.
A client application may want to replace the standard C99
math functions with Sleef's implementations by creating its
own symbolic aliases.  That technique can only work if Sleef's
functions match the calling convention of the functions being
replaced.  On Windows, everybody either uses Microsoft's C
run-time library or rarely, an open-source one, and in both
cases the standard C library functions are defined as __cdecl
not __stdcall.