Fix CMake failing to re-generate sleef.h correctly on re-building
The configure_file(...) command in CMake only executes once.
Then, generated contents get appended to the output file.

If the output file needs to be re-generated, the new contents are
appended to the existing contents, which leads to a completely
failing build.  The new contents need to replace the old contents.
Do that by stashing the output of configure_file(...) into a
different file (which does not get re-generated), and then
copy the contents of that file to the beginning of the sleef.h
generated header.