Add option to disable auto-linking on Windows
Some clients might want to rename the import library,
or otherwise manage their linker dependencies explicitly,
and the hard-coded name in the auto-linking prevents that.

This option is triggered by defining SLEEF_DISABLE_AUTOLINK.
There is already another macro SLEEF_STATIC_LIBS that clients
are expected to define if they link to Sleef statically.

This author would prefer, for his own use, for the default
to be opposite, to disable auto-linking, but that cannot be
changed without breaking backwards compatibility for other
users, so it is not done.